app/requirements.txt
pyrogram
flask
pymupdf
python-dotenv
TgCrypto